Move project help content into its own component
